Do not cache the item's parent category. Closes #104
Lithium's model classes are not the entity, so setting things to $this have terrible side effects. Removing the caching here might not be ideal, but it fixes the bug. If we find out that we need the caching again, feel free to add it, but add it *to the entity*, not the Item's class.