Prevent fatal error on events feed update, Closes #118
Add missing namespace to properly cath feed exceptions in the UpdateEventsService
and disable the extesion if an error occur to prevent the same error on every feed update.
But this is a silent error and the site administrator is not warned about the error or of the extension disabling. What must be done in the next releases.