Create the service ItemCreation
The ItemCreation service is responsible for items creation and jobs schedules to prevent different
behaviors when creating items and remove duplicated code.

It's responsible for schedule items push notifications to
make sure that the notifications are sent only when creating an item. Also prevents
send notification whem the item is created from a first feed import.

The service also schedule the job to set medias file size if the created item has any media.

It's also responsible for set the sort order of the created item.