Validates item types and site and clean item creation code, closes #117
Validate if the item type and site is equal to the parent category, to prevent
invalid the creation or update of an invalid category.

Improves the error messages of API and Web Interface when creating an item to
allow the user know what fields of the item are invalid.

Also sets the default locale of the I18n class to enable translation in the API
without more configurations.