Prevent send push notification if no device available, closes #151
If a site not have any visitor that already logged in, the push
notification is sent to all visitors because the devices filter in
the request sent to pushwoosh is not set.

To prevent this an push notification must be sent only if the devices
tokens are specified in the Api request.